# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BNDSphere API",
    "description": "Backend Service for BNDSphere",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/uploads/initiate": {
      "post": {
        "tags": ["Uploads"],
        "summary": "Initiate Upload",
        "operationId": "initiate_upload_api_v1_uploads_initiate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/InitiateUploadRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitiateUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/uploads/confirm": {
      "post": {
        "tags": ["Uploads"],
        "summary": "Confirm Upload",
        "operationId": "confirm_upload_api_v1_uploads_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ConfirmUploadRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfirmUploadRequest": {
        "properties": {
          "scene": { "$ref": "#/components/schemas/UploadScene" },
          "object_key": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Object Key"
          }
        },
        "type": "object",
        "required": ["scene", "object_key"],
        "title": "ConfirmUploadRequest"
      },
      "ConfirmUploadResponse": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          }
        },
        "type": "object",
        "required": ["url"],
        "title": "ConfirmUploadResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InitiateUploadRequest": {
        "properties": {
          "scene": { "$ref": "#/components/schemas/UploadScene" },
          "filename": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Filename"
          },
          "content_type": { "type": "string", "title": "Content Type" },
          "size": { "type": "integer", "exclusiveMinimum": 0, "title": "Size" }
        },
        "type": "object",
        "required": ["scene", "filename", "content_type", "size"],
        "title": "InitiateUploadRequest"
      },
      "InitiateUploadResponse": {
        "properties": {
          "object_key": { "type": "string", "title": "Object Key" },
          "upload_url": { "type": "string", "title": "Upload Url" },
          "expires_seconds": { "type": "integer", "title": "Expires Seconds" }
        },
        "type": "object",
        "required": ["object_key", "upload_url", "expires_seconds"],
        "title": "InitiateUploadResponse"
      },
...
      "UploadScene": {
        "type": "string",
        "enum": ["avatar", "club_logo", "activity_poster", "application_file"],
        "title": "UploadScene"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": { "scopes": {}, "tokenUrl": "/api/v1/auth/login" }
        }
      }
    }
  }
}